+++ crates/wasmtime/src/store.rs
@@ -95,7 +95,8 @@ use std::task::{Context, Poll};
 use wasmtime_runtime::{
     InstanceAllocationRequest, InstanceAllocator, InstanceHandle, ModuleInfo,
     OnDemandInstanceAllocator, SignalHandler, StorePtr, VMCallerCheckedAnyfunc, VMContext,
-    VMExternRef, VMExternRefActivationsTable, VMInterrupts, VMSharedSignatureIndex, VMTrampoline,
+    VMExternRef, VMExternRefActivationsTable, VMRuntimeLimits, VMSharedSignatureIndex,
+    VMTrampoline,
 };
 
 mod context;
@@ -255,7 +256,7 @@ pub struct StoreOpaque {
     _marker: marker::PhantomPinned,
 
     engine: Engine,
-    interrupts: Arc<VMInterrupts>,
+    runtime_limits: VMRuntimeLimits,
     instances: Vec<StoreInstance>,
     signal_handler: Option<Box<SignalHandler<'static>>>,
     externref_activations_table: VMExternRefActivationsTable,
@@ -273,7 +274,7 @@ pub struct StoreOpaque {
     memory_limit: usize,
     table_count: usize,
     table_limit: usize,
-    /// An adjustment to add to the fuel consumed value in `interrupts` above
+    /// An adjustment to add to the fuel consumed value in `runtime_limits` above
     /// to get the true amount of fuel consumed.
     fuel_adj: i64,
     #[cfg(feature = "async")]
@@ -434,7 +435,7 @@ impl<T> Store<T> {
             inner: StoreOpaque {
                 _marker: marker::PhantomPinned,
                 engine: engine.clone(),
-                interrupts: Default::default(),
+                runtime_limits: Default::default(),
                 instances: Vec::new(),
                 signal_handler: None,
                 externref_activations_table: VMExternRefActivationsTable::new(),
@@ -626,89 +627,6 @@ impl<T> Store<T> {
     /// Returns the [`Engine`] that this store is associated with.
     pub fn engine(&self) -> &Engine {
         self.inner.engine()
-    }
-
-    /// Creates an [`InterruptHandle`] which can be used to interrupt the
-    /// execution of instances within this `Store`.
-    ///
-    /// An [`InterruptHandle`] handle is a mechanism of ensuring that guest code
-    /// doesn't execute for too long. For example it's used to prevent wasm
-    /// programs for executing infinitely in infinite loops or recursive call
-    /// chains.
-    ///
-    /// The [`InterruptHandle`] type is sendable to other threads so you can
-    /// interact with it even while the thread with this `Store` is executing
-    /// wasm code.
-    ///
-    /// There's one method on an interrupt handle:
-    /// [`InterruptHandle::interrupt`]. This method is used to generate an
-    /// interrupt and cause wasm code to exit "soon".
-    ///
-    /// ## When are interrupts delivered?
-    ///
-    /// The term "interrupt" here refers to one of two different behaviors that
-    /// are interrupted in wasm:
-    ///
-    /// * The head of every loop in wasm has a check to see if it's interrupted.
-    /// * The prologue of every function has a check to see if it's interrupted.
-    ///
-    /// This interrupt mechanism makes no attempt to signal interrupts to
-    /// native code. For example if a host function is blocked, then sending
-    /// an interrupt will not interrupt that operation.
-    ///
-    /// Interrupts are consumed as soon as possible when wasm itself starts
-    /// executing. This means that if you interrupt wasm code then it basically
-    /// guarantees that the next time wasm is executing on the target thread it
-    /// will return quickly (either normally if it were already in the process
-    /// of returning or with a trap from the interrupt). Once an interrupt
-    /// trap is generated then an interrupt is consumed, and further execution
-    /// will not be interrupted (unless another interrupt is set).
-    ///
-    /// When implementing interrupts you'll want to ensure that the delivery of
-    /// interrupts into wasm code is also handled in your host imports and
-    /// functionality. Host functions need to either execute for bounded amounts
-    /// of time or you'll need to arrange for them to be interrupted as well.
-    ///
-    /// ## Return Value
-    ///
-    /// This function returns a `Result` since interrupts are not always
-    /// enabled. Interrupts are enabled via the
-    /// [`Config::interruptable`](crate::Config::interruptable) method, and if
-    /// this store's [`Config`](crate::Config) hasn't been configured to enable
-    /// interrupts then an error is returned.
-    ///
-    /// ## Examples
-    ///
-    /// ```
-    /// # use anyhow::Result;
-    /// # use wasmtime::*;
-    /// # fn main() -> Result<()> {
-    /// // Enable interruptable code via `Config` and then create an interrupt
-    /// // handle which we'll use later to interrupt running code.
-    /// let engine = Engine::new(Config::new().interruptable(true))?;
-    /// let mut store = Store::new(&engine, ());
-    /// let interrupt_handle = store.interrupt_handle()?;
-    ///
-    /// // Compile and instantiate a small example with an infinite loop.
-    /// let module = Module::new(&engine, r#"
-    ///     (func (export "run") (loop br 0))
-    /// "#)?;
-    /// let instance = Instance::new(&mut store, &module, &[])?;
-    /// let run = instance.get_typed_func::<(), (), _>(&mut store, "run")?;
-    ///
-    /// // Spin up a thread to send us an interrupt in a second
-    /// std::thread::spawn(move || {
-    ///     std::thread::sleep(std::time::Duration::from_secs(1));
-    ///     interrupt_handle.interrupt();
-    /// });
-    ///
-    /// let trap = run.call(&mut store, ()).unwrap_err();
-    /// assert!(trap.to_string().contains("wasm trap: interrupt"));
-    /// # Ok(())
-    /// # }
-    /// ```
-    pub fn interrupt_handle(&self) -> Result<InterruptHandle> {
-        self.inner.interrupt_handle()
     }
 
     /// Perform garbage collection of `ExternRef`s.
@@ -921,13 +839,6 @@ impl<'a, T> StoreContext<'a, T> {
         self.0.engine()
     }
 
-    /// Returns an [`InterruptHandle`] to interrupt wasm execution.
-    ///
-    /// See [`Store::interrupt_handle`] for more information.
-    pub fn interrupt_handle(&self) -> Result<InterruptHandle> {
-        self.0.interrupt_handle()
-    }
-
     /// Access the underlying data owned by this `Store`.
     ///
     /// Same as [`Store::data`].
@@ -961,13 +872,6 @@ impl<'a, T> StoreContextMut<'a, T> {
     /// Returns the underlying [`Engine`] this store is connected to.
     pub fn engine(&self) -> &Engine {
         self.0.engine()
-    }
-
-    /// Returns an [`InterruptHandle`] to interrupt wasm execution.
-    ///
-    /// See [`Store::interrupt_handle`] for more information.
-    pub fn interrupt_handle(&self) -> Result<InterruptHandle> {
-        self.0.interrupt_handle()
     }
 
     /// Perform garbage collection of `ExternRef`s.
@@ -1111,16 +1015,6 @@ impl StoreOpaque {
         &mut self.store_data
     }
 
-    pub fn interrupt_handle(&self) -> Result<InterruptHandle> {
-        if self.engine.config().tunables.interruptable {
-            Ok(InterruptHandle {
-                interrupts: self.interrupts.clone(),
-            })
-        } else {
-            bail!("interrupts aren't enabled for this `Store`")
-        }
-    }
-
     #[inline]
     pub(crate) fn modules_mut(&mut self) -> &mut ModuleRegistry {
         &mut self.modules
@@ -1148,8 +1042,8 @@ impl StoreOpaque {
     }
 
     #[inline]
-    pub fn interrupts(&self) -> &VMInterrupts {
-        &self.interrupts
+    pub fn runtime_limits(&self) -> &VMRuntimeLimits {
+        &self.runtime_limits
     }
 
     #[inline]
@@ -1263,7 +1157,7 @@ impl StoreOpaque {
         if !self.engine.config().tunables.consume_fuel {
             return None;
         }
-        let consumed = unsafe { *self.interrupts.fuel_consumed.get() };
+        let consumed = unsafe { *self.runtime_limits.fuel_consumed.get() };
         Some(u64::try_from(self.fuel_adj + consumed).unwrap())
     }
 
@@ -1335,7 +1229,7 @@ impl StoreOpaque {
         // reasonable amount of time anyway.
         let fuel = i64::try_from(fuel).unwrap_or(i64::max_value());
         let adj = self.fuel_adj;
-        let consumed_ptr = unsafe { &mut *self.interrupts.fuel_consumed.get() };
+        let consumed_ptr = unsafe { &mut *self.runtime_limits.fuel_consumed.get() };
 
         match (consumed_ptr.checked_sub(fuel), adj.checked_add(fuel)) {
             // If we succesfully did arithmetic without overflowing then we can
@@ -1358,7 +1252,7 @@ impl StoreOpaque {
     }
 
     fn consume_fuel(&mut self, fuel: u64) -> Result<u64> {
-        let consumed_ptr = unsafe { &mut *self.interrupts.fuel_consumed.get() };
+        let consumed_ptr = unsafe { &mut *self.runtime_limits.fuel_consumed.get() };
         match i64::try_from(fuel)
             .ok()
             .and_then(|fuel| consumed_ptr.checked_add(fuel))
@@ -1394,8 +1288,8 @@ impl StoreOpaque {
     }
 
     #[inline]
-    pub fn vminterrupts(&self) -> *mut VMInterrupts {
-        &*self.interrupts as *const VMInterrupts as *mut VMInterrupts
+    pub fn vmruntime_limits(&self) -> *mut VMRuntimeLimits {
+        &self.runtime_limits as *const VMRuntimeLimits as *mut VMRuntimeLimits
     }
 
     pub unsafe fn insert_vmexternref_without_gc(&mut self, r: VMExternRef) {
@@ -1731,8 +1625,8 @@ impl AsyncCx {
 }
 
 unsafe impl<T> wasmtime_runtime::Store for StoreInner<T> {
-    fn vminterrupts(&self) -> *mut VMInterrupts {
-        <StoreOpaque>::vminterrupts(self)
+    fn vmruntime_limits(&self) -> *mut VMRuntimeLimits {
+        <StoreOpaque>::vmruntime_limits(self)
     }
 
     fn epoch_ptr(&self) -> *const AtomicU64 {
@@ -1912,13 +1806,13 @@ impl<T> StoreInner<T> {
         // Set a new deadline based on the "epoch deadline delta".
         //
         // Safety: this is safe because the epoch deadline in the
-        // `VMInterrupts` is accessed only here and by Wasm guest code
+        // `VMRuntimeLimits` is accessed only here and by Wasm guest code
         // running in this store, and we have a `&mut self` here.
         //
         // Also, note that when this update is performed while Wasm is
         // on the stack, the Wasm will reload the new value once we
         // return into it.
-        let epoch_deadline = unsafe { (*self.vminterrupts()).epoch_deadline.get_mut() };
+        let epoch_deadline = unsafe { (*self.vmruntime_limits()).epoch_deadline.get_mut() };
         *epoch_deadline = self.engine().current_epoch() + delta;
     }
 
@@ -1926,7 +1820,7 @@ impl<T> StoreInner<T> {
         // Safety: this is safe because, as above, it is only invoked
         // from within `new_epoch` which is called from guest Wasm
         // code, which will have an exclusive borrow on the Store.
-        let epoch_deadline = unsafe { (*self.vminterrupts()).epoch_deadline.get_mut() };
+        let epoch_deadline = unsafe { (*self.vmruntime_limits()).epoch_deadline.get_mut() };
         *epoch_deadline
     }
 }
@@ -1980,28 +1874,6 @@ impl Drop for StoreOpaque {
 impl wasmtime_runtime::ModuleInfoLookup for ModuleRegistry {
     fn lookup(&self, pc: usize) -> Option<Arc<dyn ModuleInfo>> {
         self.lookup_module(pc)
-    }
-}
-
-/// A threadsafe handle used to interrupt instances executing within a
-/// particular `Store`.
-///
-/// This structure is created by the [`Store::interrupt_handle`] method.
-#[derive(Debug)]
-pub struct InterruptHandle {
-    interrupts: Arc<VMInterrupts>,
-}
-
-impl InterruptHandle {
-    /// Flags that execution within this handle's original [`Store`] should be
-    /// interrupted.
-    ///
-    /// This will not immediately interrupt execution of wasm modules, but
-    /// rather it will interrupt wasm execution of loop headers and wasm
-    /// execution of function entries. For more information see
-    /// [`Store::interrupt_handle`].
-    pub fn interrupt(&self) {
-        self.interrupts.interrupt()
     }
 }